+++ R/plot_tx_context.R
@@ -12,8 +12,7 @@
 #' @export
 plot_tx_context <- function(
     exon_table,
-    custom_colors = NULL
-  ) {
+    custom_colors = NULL) {
   if (isFALSE(tibble::is_tibble(exon_table))) {
     exon_table <- tibble::as_tibble(exon_table)
   }