+++ R/prepare_profile_data.R
@@ -1,5 +1,62 @@
-#' Prepare Data for Profile Plot
-#' @return a `tibble` with data to plot.
+#' Prepare Data for Gene and Transcript Expression Profile Plot
+#'
+#' This function processes gene and transcript-level expression data, along with differential
+#' expression results, to prepare a tidy data frame suitable for plotting expression profiles
+#' across different sample groups.
+#'
+#' @param txi_gene A `tibble` or `tximport` output containing gene-level expression abundances.
+#' If `NULL`, gene-level abundances will be summarized from `txi_transcript`. Default is `NULL`.
+#' @param txi_transcript A `tibble` or `tximport` output containing transcript-level expression abundances.
+#' @param sample_metadata A `data.frame` or `tibble` containing sample metadata. The first column
+#' should contain sample names matching the column names in `txi_gene` and `txi_transcript`.
+#' @param tx_to_gene A `data.frame` or `tibble` containing transcript-to-gene mapping information.
+#' Must include columns specified by `gene_col` and `tx_col`.
+#' @param de_result_gene A `data.frame` or `tibble` containing differential expression results at the gene level.
+#' Must include `gene_name`, `log2FC`, and `qvalue` columns.
+#' @param de_result_transcript A `data.frame` or `tibble` containing differential expression results at the transcript level.
+#' Must include `transcript_name`, `log2FC`, and `qvalue` columns.
+#' @param var A string specifying the column name in `sample_metadata` that indicates the grouping variable (e.g., treatment, condition).
+#' @param var_levels A character vector specifying the levels of `var` to include in the contrasts.
+#' @param gene_col A string specifying the column name in `tx_to_gene` that contains gene names. Default is `"gene_name"`.
+#' @param tx_col A string specifying the column name in `tx_to_gene` that contains transcript names. Default is `"transcript_name"`.
+#' @param pvalue_cutoff A numeric value specifying the p-value cutoff for determining significant differential expression. Default is `0.05`.
+#' @param lfc_cutoff A numeric value specifying the log2 fold-change cutoff for determining significant differential expression. Default is `1`.
+#' @param use_fdr A logical value indicating whether to use the false discovery rate (`qvalue`) instead of p-value for significance cutoff. Default is `TRUE`.
+#'
+#' @return A `tibble` containing processed expression data and differential expression flags, ready for plotting.
+#'
+#' @details The function combines gene and transcript expression data with differential expression results to generate a tidy data frame. It filters significant genes and transcripts based on specified cutoffs and prepares the data for plotting expression profiles across specified sample groups.
+#'
+#' @examples
+#' # Assuming txi_gene, txi_transcript, sample_metadata, tx_to_gene, de_result_gene,
+#' # and de_result_transcript are pre-loaded data frames:
+#'
+#' # Prepare data for plotting
+#' expr_df <- prepare_profile_data(
+#'   txi_gene = txi_gene,
+#'   txi_transcript = txi_transcript,
+#'   sample_metadata = sample_metadata,
+#'   tx_to_gene = tx_to_gene,
+#'   de_result_gene = de_result_gene,
+#'   de_result_transcript = de_result_transcript,
+#'   var = "condition",
+#'   var_levels = c("control", "treatment"),
+#'   gene_col = "gene_name",
+#'   tx_col = "transcript_name",
+#'   pvalue_cutoff = 0.05,
+#'   lfc_cutoff = 1,
+#'   use_fdr = TRUE
+#' )
+#'
+#' # View the prepared data
+#' head(expr_df)
+#'
+#' # Plotting example (assuming ggplot2 is installed)
+#' library(ggplot2)
+#' ggplot(expr_df, aes(x = condition, y = mean_TPM, fill = DE)) +
+#'   geom_bar(stat = "identity", position = position_dodge()) +
+#'   facet_wrap(~ parent_gene + transcript_type)
+#' 
 #' @export
 prepare_profile_data <- function(
     txi_gene = NULL, # txi abundance genes
@@ -119,7 +176,7 @@ prepare_profile_data <- function(
     tidyr::pivot_longer(-genename, names_to = "sample", values_to = "TPM") |>
     dplyr::left_join(condition_df, by = c("sample" = sample_col)) |>
     dplyr::group_by(genename, .data[[var]]) |>
-    dplyr::summarise(mean_TPM = mean(TPM), SD = sd(TPM)) |>
+    dplyr::summarise(mean_TPM = base::mean(TPM), SD = stats::sd(TPM)) |>
     dplyr::ungroup()
 
   # Add differential expression information
@@ -144,7 +201,7 @@ prepare_profile_data <- function(
     tidyr::pivot_longer(-genename, names_to = "sample", values_to = "TPM") |>
     dplyr::left_join(condition_df, by = c("sample" = sample_col)) |>
     dplyr::group_by(genename, .data[[var]]) |>
-    dplyr::summarise(mean_TPM = mean(TPM), SD = sd(TPM)) |>
+    dplyr::summarise(mean_TPM = base::mean(TPM), SD = stats::sd(TPM)) |>
     dplyr::ungroup()
 
   de_tx_df <- de_result_transcript |>
@@ -209,7 +266,17 @@ summarize_to_gene <- function(txi_transcript, tx_to_gene) {
   return(txi_gene)
 }
 
-# Convert `matrix` and `data.frame` to `tibble`
+
+#' Summarize Transcript-level Expression to Gene-level
+#'
+#' This function aggregates transcript-level expression data to gene-level by calculating the mean expression of transcripts belonging to the same gene.
+#'
+#' @param txi_transcript A `tibble` containing transcript-level expression abundances.
+#' @param tx_to_gene A `data.frame` or `tibble` containing transcript-to-gene mapping information, including `transcript_id` and `gene_id` columns.
+#'
+#' @return A `tibble` containing gene-level expression abundances.
+#'
+#' @keywords internal
 convert_to_isoformic_tibble <- function(txi_transcript) {
   if (isFALSE(inherits(x = txi_transcript, what = c("tbl_df")))) {
     if (isTRUE(inherits(x = txi_transcript, what = c("matrix"))) || isTRUE(inherits(x = txi_transcript, what = c("data.frame")))) {