+++ R/make_tx_to_gene.R
@@ -15,6 +15,7 @@
 #' For GTF or GFF3 files, support is not yet implemented.
 #'
 #' @examples
+#' \dontrun{
 #' # Assuming you have downloaded the GENCODE transcriptome FASTA file:
 #' fasta_file <- download_reference(
 #'   version = "43",
@@ -28,9 +29,13 @@
 #'
 #' # View the first few rows
 #' head(tx_to_gene)
+#' }
 #'
 #' @export
-make_tx_to_gene <- function(file_path, file_type = c("fasta", "gff", "gtf")) {
+make_tx_to_gene <- function(
+    file_path,
+    file_type = c("fasta", "gff", "gtf")) {
+  .data <- rlang::.data
   file_type <- rlang::arg_match(file_type)
   if (!isTRUE(fs::file_exists(file_path))) {
     cli::cli_abort(
@@ -44,14 +49,14 @@ make_tx_to_gene <- function(file_path, file_type = c("fasta", "gff", "gtf")) {
     header_table <- header_table |>
       tibble::as_tibble() |>
       tidyr::separate(
-        value,
+        .data[["value"]],
         into = paste0("col_", 1:9),
         sep = "\\|"
       ) |>
       dplyr::select(-c("col_9"))
     header_table$col_1 <- stringr::str_remove(header_table$col_1, ">")
   } else if (isTRUE(file_type == "gff")) {
-    # TODO: Actually implement parsing GFF3 and GTF files
+    # TODO: @luciorq Actually implement parsing GFF3 and GTF files
     cli::cli_abort(
       c(
         `x` = "{.var GTF} and {.var GFF3} files are not supported yet.",