+++ NEWS.md
@@ -1,3 +1,7 @@
+# isoformic (Development Version)
+
+* Nothing yet
+
 # isoformic v0.1.1
 
 ## Reference and Annotation