+++ NEWS.md
@@ -1,4 +1,4 @@
-## Isoformic [Unreleased]
+## Isoformic [0.1.2] - 2025-10-06
 
 ### Added
 
@@ -35,7 +35,7 @@
 
 * Original workflow style code added.
 
-[unreleased]: https://github.com/luciorq/isoformic/compare/v0.1.1...HEAD
+[0.1.2]: <https://github.com/luciorq/isoformic/compare/v0.1.1...v0.1.2>
 [0.1.1]: <https://github.com/luciorq/isoformic/compare/v0.1.0...v0.1.1>
 [0.1.0]: <https://github.com/luciorq/isoformic/compare/v0.0.1...v0.1.0>
 [0.0.1]: <https://github.com/luciorq/isoformic/releases/tag/v0.0.1>